# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PhoneMail API",
    "version": "0.1.0",
...
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
...
    "/api/auth/register": { "post": { "parameters": [{ "$ref": "#/components/parameters/AuthTransport" }], "requestBody": { "$ref": "#/components/requestBodies/PasswordRegistration" }, "responses": { "201": { "$ref": "#/components/responses/Auth" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" } } } },
    "/api/auth/login": { "post": { "parameters": [{ "$ref": "#/components/parameters/AuthTransport" }], "requestBody": { "$ref": "#/components/requestBodies/Auth" }, "responses": { "200": { "$ref": "#/components/responses/Auth" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" } } } },
    "/api/auth/otp/register": { "post": { "parameters": [{ "$ref": "#/components/parameters/AuthTransport" }], "requestBody": { "$ref": "#/components/requestBodies/OtpRegistration" }, "responses": { "201": { "$ref": "#/components/responses/Auth" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/api/auth/otp/login": { "post": { "parameters": [{ "$ref": "#/components/parameters/AuthTransport" }], "requestBody": { "$ref": "#/components/requestBodies/OtpLogin" }, "responses": { "200": { "$ref": "#/components/responses/Auth" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" } } } },
    "/api/auth/phone-change": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/AuthTransport" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "requestBody": { "$ref": "#/components/requestBodies/PhoneChange" }, "responses": { "200": { "$ref": "#/components/responses/Auth" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/api/auth/refresh": { "post": { "security": [{ "refreshCookie": [] }, {}], "description": "Rotates an opaque bearer refreshToken supplied in the JSON body or the HttpOnly phonemail_refresh cookie. Use exactly one transport; cookie transport requires the CSRF cookie value in X-CSRF-Token. Access credentials expire after 15 minutes and refresh sessions after 30 days. Repeating the immediately previous token within 90 seconds returns the same encrypted cached response; later replay revokes only that session family.", "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }], "requestBody": { "required": false, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "properties": { "refreshToken": { "type": "string", "pattern": "^[A-Za-z0-9_-]{43}$" } } } } } }, "responses": { "200": { "$ref": "#/components/responses/Auth" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
    "/api/auth/logout": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "responses": { "204": { "description": "Revokes the current session family and clears access, refresh, and CSRF cookies" }, "401": { "$ref": "#/components/responses/Error" } } } },
    "/api/auth/me": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "responses": { "200": { "description": "Current active account profile" }, "401": { "$ref": "#/components/responses/Error" } } } },
...
    "/api/me": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "responses": { "200": { "description": "Authenticated account personal details, preferences-safe profile and owned addresses" }, "401": { "$ref": "#/components/responses/Error" } } }, "patch": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "requestBody": { "$ref": "#/components/requestBodies/ProfilePatch" }, "responses": { "200": { "description": "Updated display name and preferred language; privileged account status and terms fields are not editable here" }, "400": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "parameters": {
      "AuthTransport": { "name": "X-Auth-Transport", "in": "header", "required": false, "description": "Set to cookie to set access, refresh, and CSRF cookies instead of returning bearer credentials", "schema": { "type": "string", "enum": ["cookie"] } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": false, "schema": { "type": "string", "minLength": 8, "maxLength": 128 } },
...
      "CsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": false, "description": "Required with the phonemail_refresh cookie; value must match the phonemail_csrf cookie", "schema": { "type": "string" } },
...
    },
    "requestBodies": {
      "Auth": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["phone", "password"], "properties": { "phone": { "type": "string", "description": "International E.164 or national form when accompanied by country" }, "country": { "type": "string", "pattern": "^[A-Z]{2}$", "description": "ISO 3166-1 alpha-2; required for national input unless PHONE_DEFAULT_COUNTRY is explicitly configured" }, "password": { "type": "string", "minLength": 6 } } } } } },
      "PasswordRegistration": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["phone", "password", "termsAccepted"], "properties": { "phone": { "type": "string" }, "country": { "type": "string", "pattern": "^[A-Z]{2}$" }, "password": { "type": "string", "minLength": 6, "description": "UTF-8 encoded size must not exceed 72 bytes" }, "displayName": { "type": "string", "maxLength": 100 }, "language": { "type": "string", "pattern": "^[a-z]{2}(-[A-Z]{2})?$" }, "termsAccepted": { "type": "boolean", "const": true }, "termsVersion": { "type": "string", "example": "mvp-1" }, "signupChannel": { "type": "string", "enum": ["mobile", "web", "portal", "ivr"] } } } } } },
...
      "OtpRegistration": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/OtpProof" },
                {
                  "type": "object",
                  "required": ["termsAccepted"],
                  "properties": {
                    "purpose": { "type": "string", "enum": ["signup"] },
                    "termsAccepted": { "type": "boolean", "const": true },
                    "termsVersion": { "type": "string", "example": "mvp-1" },
                    "displayName": { "type": "string", "maxLength": 100 },
                    "language": { "type": "string", "pattern": "^[a-z]{2}(-[A-Z]{2})?$" },
                    "signupChannel": { "type": "string", "enum": ["mobile", "web", "portal", "ivr"] }
                  }
                }
              ]
            }
          }
        }
      },
      "OtpLogin": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["phone", "challengeId", "purpose", "code"], "properties": { "phone": { "type": "string" }, "country": { "type": "string", "pattern": "^[A-Z]{2}$" }, "challengeId": { "type": "string", "format": "uuid" }, "purpose": { "type": "string", "enum": ["login"] }, "code": { "type": "string", "pattern": "^[0-9]{6}$" } } } } } },
      "PhoneChange": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["newPhone", "challengeId", "code"], "properties": { "currentPassword": { "type": "string" }, "newPhone": { "type": "string" }, "country": { "type": "string", "pattern": "^[A-Z]{2}$" }, "challengeId": { "type": "string", "format": "uuid" }, "code": { "type": "string", "pattern": "^[0-9]{6}$" }, "oldChallengeId": { "type": "string", "format": "uuid" }, "oldCode": { "type": "string", "pattern": "^[0-9]{6}$" } } } } } },
...
      "ProfilePatch": { "required": true, "content": { "application/json": { "schema": { "type": "object", "minProperties": 1, "maxProperties": 2, "additionalProperties": false, "properties": { "displayName": { "type": "string", "maxLength": 100 }, "language": { "type": "string", "pattern": "^[a-z]{2}(-[A-Z]{2})?$" } } } } } },
...
    },
    "schemas": {
      "OtpProof": { "type": "object", "required": ["phone", "challengeId", "purpose", "code"], "properties": { "phone": { "type": "string" }, "country": { "type": "string", "pattern": "^[A-Z]{2}$" }, "challengeId": { "type": "string", "format": "uuid" }, "purpose": { "type": "string" }, "code": { "type": "string", "pattern": "^[0-9]{6}$" } } },
...
      "AuthResult": { "type": "object", "required": ["user", "token", "refreshToken"], "properties": { "user": { "$ref": "#/components/schemas/User" }, "token": { "type": "string", "description": "15-minute access token; returned for bearer transport" }, "refreshToken": { "type": "string", "description": "Opaque 30-day rotating credential; returned for bearer transport only" } } },
      "CookieAuthResult": { "type": "object", "required": ["user"], "properties": { "user": { "$ref": "#/components/schemas/User" } } },
      "User": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "phone": { "type": "string" }, "phoneE164": { "type": "string", "nullable": true }, "phoneCountry": { "type": "string", "nullable": true }, "phoneIdentityNeedsCountry": { "type": "boolean" }, "email": { "type": "string", "format": "email" }, "displayName": { "type": "string", "nullable": true }, "language": { "type": "string" }, "signupChannel": { "type": "string" }, "termsAcceptedAt": { "type": "string", "format": "date-time", "nullable": true }, "termsVersion": { "type": "string", "nullable": true } } },
      "ErrorEnvelope": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message", "requestId", "retryable"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "requestId": { "type": "string" }, "retryable": { "type": "boolean" }, "fields": { "type": "object", "additionalProperties": true } } } } },
...
    },
    "responses": {
      "Auth": { "description": "Bearer transport returns access and refresh credentials; cookie transport sets HttpOnly access/refresh cookies and returns only the user", "headers": { "X-Request-ID": { "schema": { "type": "string" }, "description": "Request correlation ID" } }, "content": { "application/json": { "schema": { "oneOf": [{ "$ref": "#/components/schemas/AuthResult" }, { "$ref": "#/components/schemas/CookieAuthResult" }] } } } },
      "Error": { "description": "Stable error envelope with code, message, requestId and retryable", "headers": { "X-Request-ID": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" }, "description": "Present for rate limits and retryable transient failures" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    },
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_session" },
      "refreshCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_refresh" }
    }
  }
}